pyproject.toml
[tool.poetry]
name = "langchain-agent"
version = "0.1.0"
description = "Agents to learn langchain"
authors = ["Fernando Suárez Jiménez <fsuarezj@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.7"
langserve = {extras = ["all"], version = "^0.2.0"}
langchain-community = "^0.2.0"
langgraph = "^0.0.55"
grandalf = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
